refactor(analysis): move decoder impl into separate module
